requirements.txt
aiohttp==2.2.5
async-timeout==2.0.1
cffi==1.11.5
chardet==3.0.4
discord.py==1.0.0a0
idna==2.6
multidict==4.2.0
pycparser==2.18
PyNaCl==1.1.2
six==1.11.0
websockets==3.4
yarl==1.1.1
